рефакторинг тз пункт 1.1.2
